added no_mmap & mlock parameters to llama.cpp
